requirements.txt
asgiref==2.1.2
async-timeout==2.0.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
channels==2.0.0
constantly==15.1.0
daphne==2.0.1
Django==2.0
django-filter==1.1.0
django-debug-toolbar==1.9.1
djangorestframework==3.7.7
hyperlink==17.3.1
incremental==17.5.0
Markdown==2.6.11
psycopg2==2.7.3.2
pytz==2017.3
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
zope.interface==4.4.3
aioredis==1.0.0
channels-redis==2.0.1
hiredis==0.2.0
msgpack==0.5.3
redis==2.10.6
click==6.7
django-rq==1.0.1
rq==0.10.0
PyDispatcher==2.0.5
asn1crypto==0.24.0
cffi==1.11.4
cryptography==2.1.4
cssselect==1.0.3
idna==2.6
lxml==4.1.1
parsel==1.3.1
pyOpenSSL==17.5.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
queuelib==1.4.2
scrapy==1.5.0
service-identity==17.0.0
w3lib==1.19.0
scrapyd==1.2.0
certifi==2018.1.18
chardet==3.0.4
python-scrapyd-api==2.0.1
requests==2.18.4
urllib3==1.22
xlsxwriter==1.0.2
django-cors-headers==2.1.0
scrapy-djangoitem==1.1.1
numpy==1.14.0
scipy==1.0.0
pandas==0.22.0
python-dateutil==2.6.1
scikit-learn==0.19.1
boto3==1.5.25
botocore==1.8.39
docutils==0.14
jmespath==0.9.3
s3transfer==0.1.12
scrapy-proxies==0.4
beautifulsoup4==4.6.0
croniter==0.3.20
rq-scheduler==0.8.1
fake-useragent==0.1.10
xlrd==1.1.0
splash==3.2
whitenoise==3.3.1
brotlipy==0.7.0
fuzzywuzzy==0.16.0
python-Levenshtein==0.12.0
zeep[async]
django-allauth==0.35.0
